enabled migration

+++ Magni.Core/Models/Student.cs
@@ -17,7 +17,7 @@ namespace Magni.Core.Models
         [Display(Name = "Birthday")]
         public DateTime Birthday { get; set; }
         public string? RegistrationNumber { get; set; }
-        public int CourseId { get; set; }
+        public long CourseId { get; set; }
         [ForeignKey("CourseId")]
         public Course? Course { get; set; }
     }
